+++ Juego/HolaSDL/Trigger.h
@@ -6,7 +6,7 @@ class Trigger :
 	public Objeto
 {
 public:
-	Trigger(Game* juego, /*Game::Texturas_t textsolo para comprobacion visual*/ int x, int y);
+	Trigger(Game* juego, /*Game::Texturas_t textsolo para comprobacion visual*/ float x, float y);
 	~Trigger(){}
 	void update(Uint32 delta);
 	Objeto_t getId() { return Objeto_t::OTrigger; }
